Updated to upstream version 1.0.3
Could not reproduce the lintina warning Frederic saw on a testing image with
pbuilder. We may need further investigation on this issue.

The database issue still remains but should not be critical. It might come from
the fact that when sphinx is importing the modules the database files are not
available.